requirements.txt
Flask
pymongo[srv]
flask-bcrypt
python-dotenv